make build pl0x

+++ src/liballoc/heap.rs
@@ -8,6 +8,7 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
+use core::isize;
 use core::num::UnsignedInt;
 #[cfg(not(test))]
 use core::ptr::PtrExt;
@@ -15,7 +16,7 @@ use core::ptr::PtrExt;
 #[inline(always)]
 fn check_size_and_alignment(size: uint, align: uint) {
     debug_assert!(size != 0);
-    debug_assert!(size <= int::MAX, "Tried to allocate too much: {} bytes", size);
+    debug_assert!(size <= isize::MAX as usize, "Tried to allocate too much: {} bytes", size);
     debug_assert!(align.is_power_of_two(), "Invalid alignment of allocation: {}", align);
 }
